[not verified] js-components: H2, H2 and Title together with Text
