--- projects/js-packages/components/components/heading/style.module.scss
@@ -1,7 +1,0 @@
-.headline-medium {
-	margin-bottom: calc( var(--spacing-base ) * 3 );
-}
-
-.headline-small {
-	margin-bottom: var( --spacing-base );
-}
