--- projects/js-packages/components/components/title/stories/style.module.scss
@@ -1,8 +1,0 @@
-.instance {
-	span {
-		color: var( --jp-gray-40 );
-		display: inline-block;
-		width: 100%;
-		background-color: var( --jp-white-off );
-	}
-}
