--- projects/js-packages/components/components/title/style.module.scss
@@ -1,4 +1,0 @@
-.title-medium,
-.title-small {
-	margin-bottom: var( --spacing-base );
-}
